agent: sysprobe-check fail silently if sysprobe configuration is missing

+++ Dockerfiles/agent/cont-init.d/60-sysprobe-check.sh
@@ -2,6 +2,10 @@
 
 sysprobe_cfg="/etc/datadog-agent/system-probe.yaml"
 
+if [[ ! -f $sysprobe_cfg ]]; then
+  exit 0
+fi
+
 if grep -Eq '^ *enable_tcp_queue_length *: *true' $sysprobe_cfg || [[ "$DD_SYSTEM_PROBE_CONFIG_ENABLE_TCP_QUEUE_LENGTH" == "true" ]]; then
   if [ -f /etc/datadog-agent/conf.d/tcp_queue_length.d/conf.yaml.example ]; then
     mv /etc/datadog-agent/conf.d/tcp_queue_length.d/conf.yaml.example \
